Apply suggestions from PR review
