+++ storage/mock/dkg_state.go
@@ -14,12 +14,12 @@ type DKGState struct {
 	mock.Mock
 }
 
-// GetDKGStarted provides a mock function with given fields: epochCounter
-func (_m *DKGState) GetDKGStarted(epochCounter uint64) (bool, error) {
+// IsDKGStarted provides a mock function with given fields: epochCounter
+func (_m *DKGState) IsDKGStarted(epochCounter uint64) (bool, error) {
 	ret := _m.Called(epochCounter)
 
 	if len(ret) == 0 {
-		panic("no return value specified for GetDKGStarted")
+		panic("no return value specified for IsDKGStarted")
 	}
 
 	var r0 bool
