+++ storage/mock/dkg_state_reader.go
@@ -14,12 +14,12 @@ type DKGStateReader struct {
 	mock.Mock
 }
 
-// GetDKGStarted provides a mock function with given fields: epochCounter
-func (_m *DKGStateReader) GetDKGStarted(epochCounter uint64) (bool, error) {
+// IsDKGStarted provides a mock function with given fields: epochCounter
+func (_m *DKGStateReader) IsDKGStarted(epochCounter uint64) (bool, error) {
 	ret := _m.Called(epochCounter)
 
 	if len(ret) == 0 {
-		panic("no return value specified for GetDKGStarted")
+		panic("no return value specified for IsDKGStarted")
 	}
 
 	var r0 bool
